+++ bcbio/pipeline/qcsummary.py
@@ -47,7 +47,7 @@ def pipeline_summary(data):
     """Provide summary information on processing sample.
     """
     work_bam = data.get("work_bam")
-    if data["sam_ref"] is not None and work_bam and has_aligned_reads(work_bam):
+    if data["sam_ref"] is not None and work_bam and work_bam.endswith(".bam") and has_aligned_reads(work_bam):
         logger.info("Generating summary files: %s" % str(data["name"]))
         data["summary"] = _run_qc_tools(work_bam, data)
     return [[data]]