+++ bcbio/upload/__init__.py
@@ -58,6 +58,7 @@ def _get_files_rnaseq(sample):
     out = _maybe_add_alignment(algorithm, sample, out)
     out = _maybe_add_counts(algorithm, sample, out)
     out = _maybe_add_cufflinks(algorithm, sample, out)
+    out = _maybe_add_assembly(algorithm, sample, out)
     out = _maybe_add_oncofuse(algorithm, sample, out)
     return _add_meta(out, sample)
 
@@ -76,7 +77,8 @@ def _add_meta(xs, sample=None, config=None):
             if isinstance(sample["name"], (tuple, list)):
                 name = sample["name"][-1]
             else:
-                name = "%s-%s" % (sample["name"], run_info.clean_name(sample["description"]))
+                name = "%s-%s" % (sample["name"],
+                                  run_info.clean_name(sample["description"]))
             x["sample"] = name
         if config:
             if "fc_name" in config and "fc_date" in config:
@@ -218,6 +220,13 @@ def _maybe_add_cufflinks(algorithm, sample, out):
                     "ext": "cufflinks"})
     return out
 
+def _maybe_add_assembly(algorithm, sample, out):
+    if "assembly" in sample:
+        out.append({"path": sample["cufflinks_dir"],
+                    "type": "directory",
+                    "ext": "assembly"})
+    return out
+
 def _has_alignment_file(algorithm, sample):
     return (((algorithm.get("aligner") or algorithm.get("realign")
               or algorithm.get("recalibrate") or algorithm.get("bam_clean")
@@ -262,5 +271,7 @@ def _get_files_project(sample, upload_config):
         out.append({"path": sample["combined_counts"]})
     if "annotated_combined_counts" in sample:
         out.append({"path": sample["annotated_combined_counts"]})
+    if "combined_fpkm" in sample:
+        out.append({"path": sample["combined_fpkm"]})
 
     return _add_meta(out, config=upload_config)